% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-createPhasingDictionary.R
\name{blessy.createPhasingDictionary}
\alias{blessy.createPhasingDictionary}
\title{Create a Phasing Dictionary for Transcripts and Domains}
\usage{
blessy.createPhasingDictionary(df, bed_df)
}
\arguments{
\item{df}{A data frame containing transcript-domain mapping, with the following required columns:
- \code{Transcript}: Transcript identifier.
- \code{DoCo}: Domain-combination string (generated by `blessy.domainPhasing`).
- \code{Gene}: Gene identifier.}

\item{bed_df}{The initial transcript annotation data frame with the following required columns:
- \code{name}: Transcript identifier.
- \code{geneName}: Gene identifier.}
}
\value{
A phasing dictionary data frame with the following columns:
  - \code{Transcript}: Transcript identifier.
  - \code{DoCo}: Domain-combination string.
  - \code{Gene}: Gene identifier.
}
\description{
This function generates a data frame summarizing gene, DoCo and transcript relationships in the annotation tracks.
}
\examples{
# Example transcript-domain data frame
phased_df <- data.frame(
  Transcript = c("tx1", "tx2"),
  DoCo = c("domainA::chr1:1000-1100(+);;; geneA", ";;; geneB"),
  Gene = c("geneA", "geneB")
)

# Example BED-like data frame
tx_df <- data.frame(
  name = c("tx1", "tx2", "tx3"),
  geneName = c("geneA", "geneB", "geneC")
)

# Create the phasing dictionary
phasing_dict <- blessy.createPhasingDictionary(phased_df, tx_df)

}
