% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-getDomainTrack.R
\name{blessy.getDomainTrack}
\alias{blessy.getDomainTrack}
\title{Fetch and Convert UCSC Domain-type Track to Data Frame}
\usage{
blessy.getDomainTrack(genome, track)
}
\arguments{
\item{genome}{A string specifying the genome/assembly identifier (e.g., "hg38").}

\item{track}{A string specifying the table identifier of the domain track (e.g., "unipDomain").}
}
\value{
A data frame in BED format containing the domain annotation track.
}
\description{
This function fetches a domain annotation track from the UCSC Genome Browser using UCSC.utils
and converts it into a BED-formatted data frame in R, notably the UniProt and Pfam tracks.
}
\examples{
# Example: Fetch the "unipDomain" annotation track
bed_df <- blessy.getDomainTrack("hg38", "unipDomain")

}
