% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-usingCustomAnnotation.R
\name{blessy.usingCustomAnnotation}
\alias{blessy.usingCustomAnnotation}
\title{blessy's Wrapper Function using Custom Annotation}
\usage{
blessy.usingCustomAnnotation(tx_df, domain_df, tx_count)
}
\arguments{
\item{tx_df}{A BED-like data frame representing transcript annotation. The annotation must include information on exon locations in `blockCount`, `blockSizes`, `blockStarts` and gene name in `geneName`.}

\item{domain_df}{A BED-like data frame representing domain annotation. The annotation must include information on domain block locations in in `blockCount`, `blockSizes`, `blockStarts`.}

\item{tx_count}{A data frame containing transcript counts with:
- \code{TranscriptID}: Transcript IDs (required as the first column).
- Additional columns representing RNA-seq counts for biological samples (must be numeric).}
}
\value{
A list containing:
  - \code{phasing_dict}: The phasing dictionary data frame
  - \code{doco_counts}: The DoCo-level count data frame
}
\description{
This wrapper function performs the workflow starting from mapping domains to transcripts
using user-provided transcript and domain annotation data frames of BED-like format. It creates a phasing dictionary
and generates DoCo-level count data frames.
}
