% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy.R
\name{blessy}
\alias{blessy}
\title{blessy: A Wrapper Function to Run All Steps in the blessy's pipeline}
\usage{
blessy(genomeAssembly, transcriptAnnotation, domainAnnotation, transcriptCount)
}
\arguments{
\item{genomeAssembly}{A string specifying the genome/assembly UCSC identifier (e.g., "hg38").}

\item{transcriptAnnotation}{A string or data frame specifying the transcript annotation track UCSC identifier (e.g., "wgEncodeGencodeBasicV44").
If a data frame is provided, refer to blessy.usingCustomAnnotation.}

\item{domainAnnotation}{A string or data frame specifying the domain annotation track.
If a data frame is provided, refer to blessy.usingCustomAnnotation.}

\item{transcriptCount}{A data frame containing transcript counts with:
- \code{TranscriptID}: Transcript IDs (required as the first column).
- Additional columns representing RNA-seq counts for biological samples (must be numeric).}
}
\value{
A list containing:
  - \code{phasing_dict}: The phasing dictionary data frame containing information of Gene, DoCo and Transcript.
  - \code{doco_counts}: The DoCo-level count data frame created by aggregating transcripts of the same DoCo.
}
\description{
This wrapper function performs the entire workflow of fetching domain and transcript tracks,
mapping domains to transcripts, deduplication, phasing, creating a phasing dictionary, 
and generating DoCo-level count data frames given a transcript count.
}
\examples{
# Example data for blessy
genomeAssembly <- "hg38"

transcriptAnnotation <- "wgEncodeGencodeBasicV44"

domainAnnotation <- "unipDomain"

transcriptCount <- data.frame(
  TranscriptID = c("tx1", "tx2", "tx3"),
  Sample1 = c(10, 20, 30),
  Sample2 = c(15, 25, 35),
  stringsAsFactors = FALSE
)

# Run blessy
results <- blessy(genomeAssembly, transcriptAnnotation, domainAnnotation, transcriptCount)

# Access the results
phasing_dict <- results$phasing_dict
doco_counts <- results$doco_counts

}
