% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blessy-usingCustomAnnotation.R
\name{blessy.usingCustomAnnotation}
\alias{blessy.usingCustomAnnotation}
\title{blessy's Wrapper Function using Custom Annotation}
\usage{
blessy.usingCustomAnnotation(
  tx_df,
  domain_df,
  tx_count,
  unique_domain = FALSE,
  coordinates = TRUE
)
}
\arguments{
\item{tx_df}{A BED-like data frame representing transcript annotation or a GRangesList object. The annotation must include information on exon locations in \code{blockCount}, \code{blockSizes}, \code{blockStarts}, and gene name in \code{geneName}.}

\item{domain_df}{A BED-like data frame representing domain annotation or a GRangesList object. The annotation must include information on domain block locations in \code{blockCount}, \code{blockSizes}, \code{blockStarts}.}

\item{tx_count}{A data frame containing transcript counts with:
- \code{TranscriptID}: Transcript IDs (required as the first column).
- Additional columns representing RNA-seq counts for biological samples (must be numeric).}

\item{unique_domain}{Logical flag indicating whether to deduplicate domains by keeping only unique domains per transcript. 
Defaults to \code{FALSE}.}

\item{coordinates}{Logical flag indicating whether to include genomic coordinates and strand information
in the \code{DoCo} string during the phasing step. Defaults to \code{TRUE}.}
}
\value{
A list containing:
  - \code{phasing_dict}: The phasing dictionary data frame.
  - \code{doco_count}: The DoCo-level count data frame.
}
\description{
This wrapper function performs the workflow starting from mapping domains to transcripts
using user-provided transcript and domain annotation data frames of BED-like format or GRangesList objects.
It creates a phasing dictionary and generates DoCo-level count data frames.
}
